requirements.txt
flask
openai-whisper
torch
numpy
waitress
sounddevice 
scipy
pyperclip
markdown